feat(contracts): move deploy related files to `/deploy`
