+++ packages/contracts/test/UserSignUpVerifier.test.ts
@@ -12,7 +12,7 @@ import {
     Reputation,
 } from './utils'
 import { SignUpProof, Unirep } from '../src'
-import { deployUnirep } from '../src/deploy'
+import { deployUnirep } from '../deploy'
 
 describe('Verify user sign up verifier', function () {
     this.timeout(30000)
